+++ test_compare/comment.lua
@@ -1,2 +1,15 @@
 -- hello
---[[world]]--[[and]]--[[you]]
+--[[world]]--[[and]]--[[you]]?12-- g'bye
+?34-- pre-if
+if e do-- in-if
+?1-- out-if
+else-- in-else
+?-1-- out-else
+end-- pre-if-2
+-- in-if-2
+?99-- out-if-2
+-- pre-if-3
+-- in-else-3
+?-99-- out-else-3
+-- in-if-4
+-- final comment (add nothing else below)