+++ test_input/comment-short.lua
@@ -1,0 +1,17 @@
+--[[this if becomes short]]
+if x then
+-- true
+a=1 --[[
+
+else incoming!
+]]else
+-- false
+a=--[[ this
+       is a really
+       small number
+]]0
+
+
+
+--end here:
+end