+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## v1.2.7c (In progress)
 * parens8 - support `dest` and `deserialize` options
 * Allow merging assignments to/from globals in safe-minify mode if there's no _ENV
+* Fix `--no-minify-comments` and support it even if minifying spaces
 
 ## v1.2.7b
 * Updated to newer version of Parens8 with reduced token/compressed/char costs.